+++ .github/workflows/deploy-pr-ephemeral-env-proxy.yaml
@@ -1,0 +1,28 @@
+name: Deploy
+on:
+  push:
+    branches:
+      - main
+      - dev
+
+env:
+  REGISTRY: ghcr.io
+  IMAGE_NAME: andrew-codes/andrew.codes
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    container:
+      image: ghcr.io/andrew-codes/gh-action-runner:latest
+    steps:
+      - name: ⬇️ Checkout source code
+        uses: actions/checkout@v3
+
+      - uses: superfly/flyctl-actions/setup-flyctl@master
+
+      - name: 🚀 Deploy emphemeral environment for pull request
+        uses: superfly/flyctl-actions@1.3
+        with:
+          args: "deploy --config fly.staging.toml --dockerfile pr.Dockerfile  --remote-only"
+        env:
+          FLY_API_TOKEN: ${{ secrets.FLY_API_TOKEN }}
