+++ .github/workflows/deploy-production.yaml
@@ -3,7 +3,6 @@ on:
   push:
     branches:
       - main
-      - dev
 
 env:
   REGISTRY: ghcr.io
@@ -19,21 +18,6 @@ jobs:
         uses: actions/checkout@v3
 
       - uses: superfly/flyctl-actions/setup-flyctl@master
-
-      - name: 👀 Read app name
-        uses: SebRollen/toml-action@v1.0.0
-        id: app_name
-        with:
-          file: "fly.toml"
-          field: "app"
-
-      - name: 🚀 Deploy Staging
-        if: ${{ github.ref == 'refs/heads/dev' }}
-        uses: superfly/flyctl-actions@1.3
-        with:
-          args: "deploy --remote-only --app ${{ steps.app_name.outputs.value }}-staging"
-        env:
-          FLY_API_TOKEN: ${{ secrets.FLY_API_TOKEN }}
 
       - name: 🚀 Deploy production
         if: ${{ github.ref == 'refs/heads/main' }}
