+++ index.js
@@ -3,7 +3,7 @@ require("@remix-run/node/dist/globals").installGlobals()
 require("dotenv/config")
 
 if (process.env.NODE_ENV === "production") {
-  require("./server-build")
+  require("./dist/app/server")
 } else {
   require("./server")
 }
