+++ package.json
@@ -11,11 +11,15 @@
     "start": "yarn node index.js",
     "start/dev": "pm2-dev ./config/pm2.config.js",
     "start/dev/remix": "remix watch",
-    "clean": "rimraf ./node_modules/.cache ./server/dist ./build ./public/build ./server-build",
+    "clean": "rimraf ./node_modules/.cache ./build ./public/build ./server-build.app ./dist",
     "prebuild": "yarn run clean",
-    "build": "concurrently 'yarn run build/server' 'yarn run build/remix'",
-    "build/remix": "remix-esbuild-override && remix build",
-    "build/server": "yarn node --require esbuild-register scripts/build-server.ts"
+    "build/app": "concurrently 'yarn run build/app/server' 'yarn run build/app/remix'",
+    "postbuild/app": "yarn node --require esbuild-register scripts/move-app-artifacts-to-dist.ts",
+    "build/app/remix": "remix-esbuild-override && remix build",
+    "build/app/server": "yarn node --require esbuild-register scripts/build-server.ts",
+    "prebuild/proxy": "yarn run clean",
+    "build/proxy": "yarn node --require esbuild-register scripts/build-proxy.ts",
+    "package/proxy": ""
   },
   "dependencies": {
     "@metronome-sh/express": "^6.0.1",
