+++ scripts/build-server.ts
@@ -5,36 +5,44 @@ import pkg from "../package.json"
 
 const here = (...s: string[]) => path.join(__dirname, ...s)
 
-const allFiles = sync("server/**/*.*", {
-  ignore: ["**/tsconfig.json", "**/eslint*", "**/__tests__/**"],
-})
+const run = async () => {
+  try {
+    const allFiles = sync("app.router/**/*.*", {
+      ignore: ["**/tsconfig.json", "**/eslint*", "**/__tests__/**"],
+    })
 
-const entries = []
-for (const file of allFiles) {
-  if (/\.(ts|js|tsx|jsx)$/.test(file)) {
-    entries.push(file)
-  } else {
-    const dest = file.replace("server", "server-build")
-    console.log(file, dest)
-    fsExtra.ensureDir(path.parse(dest).dir)
-    fsExtra.copySync(file, dest)
-    console.log(`copied: ${file.replace(`${here("server")}/`, "")}`)
+    const entries = []
+    for (const file of allFiles) {
+      if (/\.(ts|js|tsx|jsx)$/.test(file)) {
+        entries.push(file)
+      } else {
+        const dest = file.replace("app.router", "server-build.app")
+        console.log(file, dest)
+        await fsExtra.ensureDir(path.parse(dest).dir)
+        fsExtra.copySync(file, dest)
+        console.log(`copied: ${file.replace(`${here("app.router")}/`, "")}`)
+      }
+    }
+
+    console.log(`
+building...`)
+
+    await require("esbuild").build({
+      entryPoints: sync("app.router/**/*.ts"),
+      outdir: here("../server-build.app"),
+      target: [`node${pkg.engines.node}`],
+      platform: "node",
+      format: "cjs",
+      logLevel: "info",
+    })
+  } catch (error: any) {
+    console.error(error)
+    process.exit(1)
   }
 }
 
-console.log(`
-building...`)
+if (require.main !== module) {
+  run()
+}
 
-require("esbuild")
-  .build({
-    entryPoints: sync("server/**/*.ts"),
-    outdir: here("../server-build"),
-    target: [`node${pkg.engines.node}`],
-    platform: "node",
-    format: "cjs",
-    logLevel: "info",
-  })
-  .catch((error: any) => {
-    console.error(error)
-    process.exit(1)
-  })
+export default run
